feat: separeted logic, email service
